Fix structure members initialization optimize out without pragma
In helper function that create graphics pipeline I replace a lot of arguments with structure and this fixed problem in function that call this helper function. Inside helper function problem with big structure that contains states info isn't fixed so I specify instance of this function as volatile and this seems to fix problem.